common.utils: automatic refactor: URL
Move `canonical_url`, `get_tlds`, `url_regex`, `https` and `noproto` to
`common.url`.
