+++ lrrbot/linkspam.py
@@ -1,6 +1,8 @@
 import asyncio
 import re
 import logging
+
+import common.url
 from common import utils
 from lrrbot import storage
 import irc.client
@@ -10,7 +12,7 @@ log = logging.getLogger('linkspam')
 class LinkSpam:
 	def __init__(self, loop):
 		self._loop = loop
-		self._re_url = loop.run_until_complete(utils.url_regex())
+		self._re_url = loop.run_until_complete(common.url.url_regex())
 		self._rules = [
 			{
 				"re": re.compile(rule["re"], re.IGNORECASE),
@@ -39,7 +41,7 @@ class LinkSpam:
 				if url is not None:
 					urls.append(url)
 					break
-		canonical_urls = yield from asyncio.gather(*map(utils.canonical_url, urls), loop=self._loop)
+		canonical_urls = yield from asyncio.gather(*map(common.url.canonical_url, urls), loop=self._loop)
 		for original_url, url_chain in zip(urls, canonical_urls):
 			for url in url_chain:
 				for rule in self._rules:
